Provide session type information to pam for improved logind setup
Setting the XDG_SESSION_TYPE variable at session initialization ensures
that pam can provide systemd_logind with a correct description of the
actual type of session in use. This can be important since some logind
behaviors depend on the type of session being used (e.g. "idle hint
control is not supported on non-graphical sessions").

This commit updates the IPC protocol to have Request::CreateSession take
an Option<String> which may contain the session type. If the value is
None, a default of "unspecified" is used. The configuration file syntax
is also updated to recognize an optional `session_type` key in both the
`default_session` and `initial_session` sections. Again, if the key is
not specified, a default of "unspecified" will be used.